pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fin-drift-engine"
version = "1.0.0"
description = "Financial Narrative Drift Detection System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.24.0",
    "requests==2.31.0",
    "openai==1.6.1",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pyyaml==6.0.1",
    "python-dotenv==1.0.0",
    "schedule==1.2.1",
    "jinja2==3.1.3",
    "yfinance>=0.2.40",
    "scipy>=1.11.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "ruff>=0.1.0"
]
visual = [
    "streamlit>=1.29.0"
]
api = [
    "flask==3.0.0",
    "newsapi-python==0.2.7"
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
